added authenticate using the login mechanism (IMAP) this mechanism sends username/passwd base64 encoded instead of clean ASCII
git-svn-id: https://svn.ic-s.nl/svn/dbmail/trunk/dbmail@244 7b491191-dbf0-0310-aff6-d879d4d69008

+++ imaputil.c
@@ -1722,13 +1722,16 @@ void close_cache()
       cached_msg.tmpdump = NULL;
     }
 
-  if (unlink(cached_msg.filename) != 0)
+  if (cached_msg.filename[0] && unlink(cached_msg.filename) != 0)
     trace(TRACE_ERROR, "close_cache(): could not unlink temporary file %s [%s]\n", 
 	  cached_msg.filename, strerror(errno));
 
-  if (unlink(cached_msg.tmpname) != 0)
+  if (cached_msg.tmpname[0] && unlink(cached_msg.tmpname) != 0)
     trace(TRACE_ERROR, "close_cache(): could not unlink temporary file %s [%s]\n", 
 	  cached_msg.tmpname, strerror(errno));
+
+  cached_msg.filename[0] = 0;
+  cached_msg.tmpname[0] = 0;
 }
 
 /*
